[NIAK-94-assign-rooms-to-recruitment] removed time selector
