+++ rezerwacje-front/src/views/RecruitmentRooms/components/SelectRecruitmentDialog/index.js
@@ -1,0 +1,1 @@
+export { default } from './SelectRecruitmentDialog';
