--- rezerwacje-front/src/views/RecruitmentRooms/components/TimeSelectionDialog/index.js
@@ -1,1 +1,0 @@
-export { default } from './TimeSelectionDialog';
